Move ViewTraits into their own file
Check for proper copying:
git diff HEAD~1 --unified=0 core/src/View/Kokkos_ViewLegacy.hpp | cut -c2- &> tmp
diff tmp core/src/View/Kokkos_ViewTraits.hpp
